Signal processing the zdevice signal

Signal processing the zdevice signal

SignalProcessing(d, rate = 256, thresh_right = c(stoptime = 2000,
  last_window = NULL), preprocess = "ma", filtering = "lgf",
  output = "rmax", preParams = list(window = 256L),
  filtParams = list(threshold = 3.5), outputParams = list(window = 256L))

Arguments

d
A two-column data.frame with the first column being time and the second being signal

preprocess
Options are 'ma' (Moving average) and 'none'

filtering
Options are 'lgf' (local Gaussian Filtering), 'gf' (Gaussian Filtering), 'mgf' (Mixed Gaussian Filtering, not implemented), and 'none'

output
Options are 'rmean' (running mean), 'rmedian' (running median), 'apnea', and 'none'

preParams
Parameters for the chosen preprocess method, as a list

filtParams
Parameters for the chosen filtering method, as a list

outputParams
Parameters for the chosen output method, as a list